+++ index.js
@@ -1,4 +1,5 @@
 
 module.exports = {
-    compile: require('./lib/compile')
+    compile: require('./lib/compile'),
+    import: require('./lib/import')
 }